+++ pkg/utils/kubeclient/volume_claim.go
@@ -18,10 +18,6 @@ package kubeclient
 
 import (
 	"context"
-	"fmt"
-
-	"github.com/fluid-cloudnative/fluid/pkg/common"
-	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -35,32 +31,5 @@ func GetPersistentVolumeClaim(client client.Client, name, namespace string) (pvc
 			Namespace: namespace,
 		},
 		pvc)
-	return
-}
-
-// GetMountInfoFromVolumeClaim gets the mountPath and type for CSI plugin
-func GetMountInfoFromVolumeClaim(client client.Client, name, namespace string) (path string, mountType string, subpath string, err error) {
-	pvc, err := GetPersistentVolumeClaim(client, name, namespace)
-	if err != nil {
-		err = errors.Wrapf(err, "failed to get persistent volume claim")
-		return
-	}
-
-	pv, err := GetPersistentVolume(client, pvc.Spec.VolumeName)
-	if err != nil {
-		err = errors.Wrapf(err, "cannot find pvc \"%s/%s\"'s bounded PV", pvc.Namespace, pvc.Name)
-		return
-	}
-
-	if pv.Spec.CSI != nil && len(pv.Spec.CSI.VolumeAttributes) > 0 {
-		path = pv.Spec.CSI.VolumeAttributes[common.VolumeAttrFluidPath]
-		mountType = pv.Spec.CSI.VolumeAttributes[common.VolumeAttrMountType]
-		subpath = pv.Spec.CSI.VolumeAttributes[common.VolumeAttrFluidSubPath]
-	} else {
-		err = fmt.Errorf("the pvc %s in %s is not created by fluid",
-			name,
-			namespace)
-	}
-
 	return
 }
