+++ pkg/webhook/plugins/fusesidecar/fuse_sidecar.go
@@ -76,7 +76,8 @@ func (p *FuseSidecar) Mutate(pod *corev1.Pod, runtimeInfos map[string]base.Runti
 		}
 		out.DeepCopyInto(pod)
 		pvcNames := kubeclient.GetPVCNamesFromPod(pod)
-		runtimeInfos, err = webhookutils.CollectRuntimeInfosFromPVCs(p.client, pvcNames, pod.Namespace, p.log)
+		runtimeInfos, err = webhookutils.CollectRuntimeInfosFromPVCs(p.client, pvcNames, pod.Namespace, p.log,
+			utils.SkipPrecheckEnable(pod.Annotations))
 		if err != nil {
 			return shouldStop, errors.Wrapf(err, "failed to collect runtime infos from PVCs %v", pvcNames)
 		}
